% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aica.r
\name{run_aica}
\alias{run_aica}
\title{Annotation with LLM for clusters or regions.}
\usage{
run_aica(
  markers,
  topN,
  api_key,
  tissuename,
  ann_type = "celltype",
  web_search = TRUE,
  model = "deepseek-reasoner"
)
}
\arguments{
\item{markers}{The output of FindAllMarkers function in Seurat.}

\item{topN}{Default is 10. The top N genes were used for anntation.}

\item{api_key}{The api key of LLM, like deepseek or chat-gpt.}

\item{tissuename}{The tissue name.}

\item{ann_type}{Defualt is region. The annotation types include region, celltype.}

\item{web_search}{Support web search.}

\item{model}{Defualt is deepseek-reasoner}
}
\value{
Annotation results.
}
\description{
Annotation with LLM for clusters or regions.
}
